+++ rocketgram/api/chat_member.py
@@ -3,6 +3,7 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
@@ -44,23 +45,53 @@ class ChatMember:
     can_promote_members: Optional[bool]
     is_member: Optional[bool]
     can_send_messages: Optional[bool]
-    can_send_media_messages: Optional[bool]
+    can_send_audios: Optional[bool]
+    can_send_documents: Optional[bool]
+    can_send_photos: Optional[bool]
+    can_send_videos: Optional[bool]
+    can_send_video_notes: Optional[bool]
+    can_send_voice_notes: Optional[bool]
     can_send_polls: Optional[bool]
     can_send_other_messages: Optional[bool]
     can_add_web_page_previews: Optional[bool]
+
+    @property
+    def can_send_media_messages(self) -> bool:
+        warnings.warn("This field is deprecated.", DeprecationWarning)
+        return self.can_send_other_messages
 
     @classmethod
     def parse(cls, data: dict) -> Optional['ChatMember']:
         if data is None:
             return None
 
-        until_date = datetime.utcfromtimestamp(data['until_date']) if 'until_date' in data else None
-
-        return cls(User.parse(data['user']), ChatMemberStatusType(data['status']), data.get('custom_title'),
-                   data.get('is_anonymous'), until_date, data.get('can_be_edited'), data.get('can_manage_chat'),
-                   data.get('can_change_info'), data.get('can_post_messages'), data.get('can_edit_messages'),
-                   data.get('can_delete_messages'), data.get('can_manage_voice_chats'), data.get('can_invite_users'),
-                   data.get('can_restrict_members'), data.get('can_pin_messages'), data.get('can_manage_topics'),
-                   data.get('can_promote_members'), data.get('is_member'), data.get('can_send_messages'),
-                   data.get('can_send_media_messages'), data.get('can_send_polls'), data.get('can_send_other_messages'),
-                   data.get('can_add_web_page_previews'))
+        return cls(
+            User.parse(data['user']),
+            ChatMemberStatusType(data['status']),
+            data.get('custom_title'),
+            data.get('is_anonymous'),
+            datetime.utcfromtimestamp(data['until_date']) if 'until_date' in data else None,
+            data.get('can_be_edited'),
+            data.get('can_manage_chat'),
+            data.get('can_change_info'),
+            data.get('can_post_messages'),
+            data.get('can_edit_messages'),
+            data.get('can_delete_messages'),
+            data.get('can_manage_video_chats'),
+            data.get('can_invite_users'),
+            data.get('can_restrict_members'),
+            data.get('can_pin_messages'),
+            data.get('can_manage_topics'),
+            data.get('can_promote_members'),
+            data.get('is_member'),
+            data.get('can_send_messages'),
+            data.get('can_send_audios'),
+            data.get('can_send_documents'),
+            data.get('can_send_photos'),
+            data.get('can_send_videos'),
+            data.get('can_send_video_notes'),
+            data.get('can_send_voice_notes'),
+            data.get('can_send_polls'),
+            data.get('can_send_other_messages'),
+            data.get('can_add_web_page_previews'),
+        )