+++ rocketgram/api/chat_permissions.py
@@ -3,6 +3,7 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from typing import Optional
 
@@ -15,7 +16,12 @@ class ChatPermissions:
     """
 
     can_send_messages: Optional[bool]
-    can_send_media_messages: Optional[bool]
+    can_send_audios: Optional[bool]
+    can_send_documents: Optional[bool]
+    can_send_photos: Optional[bool]
+    can_send_videos: Optional[bool]
+    can_send_video_notes: Optional[bool]
+    can_send_voice_notes: Optional[bool]
     can_send_polls: Optional[bool]
     can_send_other_messages: Optional[bool]
     can_add_web_page_previews: Optional[bool]
@@ -24,12 +30,29 @@ class ChatPermissions:
     can_pin_messages: Optional[bool]
     can_manage_topics: Optional[bool]
 
+    @property
+    def can_send_media_messages(self) -> bool:
+        warnings.warn("This field is deprecated.", DeprecationWarning)
+        return self.can_send_other_messages
+
     @classmethod
     def parse(cls, data: dict) -> Optional['ChatPermissions']:
         if data is None:
             return None
 
-        return cls(data.get('can_send_messages'), data.get('can_send_media_messages'), data.get('can_send_polls'),
-                   data.get('can_send_other_messages'), data.get('can_add_web_page_previews'),
-                   data.get('can_change_info'), data.get('can_invite_users'), data.get('can_pin_messages'),
-                   data.get('can_manage_topics'))
+        return cls(
+            data.get('can_send_messages'),
+            data.get('can_send_audios'),
+            data.get('can_send_documents'),
+            data.get('can_send_photos'),
+            data.get('can_send_videos'),
+            data.get('can_send_video_notes'),
+            data.get('can_send_voice_notes'),
+            data.get('can_send_polls'),
+            data.get('can_send_other_messages'),
+            data.get('can_add_web_page_previews'),
+            data.get('can_change_info'),
+            data.get('can_invite_users'),
+            data.get('can_pin_messages'),
+            data.get('can_manage_topics')
+        )