+++ rocketgram/api/keyboard_button.py
@@ -7,6 +7,8 @@ from dataclasses import dataclass
 from typing import Optional
 
 from .keyboard_button_poll_type import KeyboardButtonPollType
+from .keyboard_button_request_chat import KeyboardButtonRequestChat
+from .keyboard_button_request_user import KeyboardButtonRequestUser
 from .web_app_info import WebAppInfo
 
 
@@ -18,6 +20,8 @@ class KeyboardButton:
     """
 
     text: str
+    request_user: Optional[KeyboardButtonRequestUser] = None
+    request_chat: Optional[KeyboardButtonRequestChat] = None
     request_contact: Optional[bool] = None
     request_location: Optional[bool] = None
     request_poll: Optional[KeyboardButtonPollType] = None